% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{celltest}
\alias{celltest}
\title{perform test if individual cells are outlying}
\usage{
celltest(x, twosided = FALSE, adjust = "fdr", ...)
}
\arguments{
\item{x}{object of class \code{polycor}}

\item{twosided}{shall test be one- or two-sided with left-tailed alternative?}

\item{adjust}{method for adjustment of p-values for multiple comparisons; default is FDR}

\item{...}{additional parameters to be passed down}
}
\description{
perform test if individual cells are outlying
}
