% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polycor_funs_helper.R
\name{init_thresholds}
\alias{init_thresholds}
\title{Initialize thresholds for K answer categories}
\usage{
init_thresholds(K)
}
\arguments{
\item{K}{number of answer categories}
}
\description{
Initialize thresholds for K answer categories
}
