% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polycor_funs_helper.R
\name{initialize_param}
\alias{initialize_param}
\title{neutral initialization of starting values}
\usage{
initialize_param(x, y)
}
\arguments{
\item{x}{vector of integer-valued responses to first item or contingency table (a \code{table} object)}

\item{y}{vector of integer-valued responses to second item; only required if \code{x} is not a contingency table}
}
\description{
neutral initialization of starting values
}
