% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{polycormat_mle}
\alias{polycormat_mle}
\title{Maximum likelihood estimation of polychoric correlation matrix}
\usage{
polycormat_mle(
  data,
  parallel = FALSE,
  num_cores = 1L,
  return_polycor = TRUE,
  variance = TRUE,
  constrained = TRUE,
  method = ifelse(constrained, "Nelder-Mead", "L-BFGS-B"),
  tol = 0.001
)
}
\arguments{
\item{data}{Data matrix of integer-valued responses, individual respondents are in rows}

\item{parallel}{Logical. Shall parallelization be used?}

\item{num_cores}{Number of cores to be used, only relevant if \code{parallel = TRUE}}

\item{return_polycor}{Logical. Shall the individual \code{polycor} objects for each item pair estimate be returned?}

\item{variance}{shall an estimated asymptotic covariance matrix be returned? Default is \code{TRUE}}

\item{constrained}{shall strict monotonicity of thresholds be explicitly enforced by linear constraints?}

\item{method}{numerical optimization method}

\item{tol}{tolerance in numerical optimization}
}
\description{
Maximum likelihood estimation of polychoric correlation matrix
}
