% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.polycor}
\alias{print.polycor}
\title{Print method for class "polycor"}
\usage{
\method{print}{polycor}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{object of class \code{polycor}}

\item{digits}{number of digits to be printed}

\item{...}{additional parameters to be passed down}
}
\description{
Print method for class "polycor"
}
