% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.robpolyserial}
\alias{print.robpolyserial}
\title{Print method for classes \code{"robpolyserial"} and \code{"polyserial"}.}
\usage{
\method{print}{robpolyserial}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{Object of class \code{"robpolyserial"} or \code{"polyserial"}.}

\item{digits}{Number of digits to be printed.}

\item{...}{Additional parameters to be passed down.}
}
\value{
A print to the console.
}
\description{
Print method for classes \code{"robpolyserial"} and \code{"polyserial"}.
}
\examples{
set.seed(123)
x <- rnorm(100)
y <- sample(c(1,2), size = 100, replace = TRUE)
fit <- polyserial(x,y) 

print(fit)
fit # equivalent

}
