\name{robord-package}
\alias{robord-package}
\alias{robord}
\docType{package}
\title{
  Robust Ordinal Data Analysis
}
\description{
  A more detailed description of what the package does. A length
  of about one to five lines is recommended.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Max Welz [aut, cre] (<https://orcid.org/0000-0003-2945-1860>)

Maintainer: Max Welz <welz@ese.eur.nl>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
