% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polycor_funs_helper.R
\name{initialize_param}
\alias{initialize_param}
\title{Neutral initialization of starting values}
\usage{
initialize_param(x, y)
}
\arguments{
\item{x}{Vector of integer-valued responses to first rating variable, or contingency table (a \code{table} object).}

\item{y}{Vector of integer-valued responses to second rating variable; only required if \code{x} is not a contingency table.}
}
\description{
Initializes starting values for numerical optimization in a neutral way. The optimization problem itself is convex, so the initialization should not matter much.
}
