% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.robpolycor}
\alias{plot.robpolycor}
\title{Plot method for class "polycor"}
\usage{
\method{plot}{robpolycor}(x, ...)
}
\arguments{
\item{x}{object of class \code{polycor}}

\item{...}{additional parameters to be passed down}
}
\description{
Plot method for class "polycor"
}
