% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polycor.R
\name{polycor}
\alias{polycor}
\title{Robust estimation of polychoric correlation coefficient}
\usage{
polycor(
  x,
  y = NULL,
  c = 1.6,
  variance = TRUE,
  constrained = TRUE,
  method = ifelse(constrained, "Nelder-Mead", "L-BFGS-B"),
  tol = 0.001,
  init = initialize_param(x, y)
)
}
\arguments{
\item{x}{vector of integer-valued responses to first item or contingency table (a \code{table} object)}

\item{y}{vector of integer-valued responses to second item; only required if \code{x} is not a contingency table}

\item{c}{tuning constant that governs robustness; eafults to 1.6}

\item{variance}{shall an estimated asymptotic covariance matrix be returned? Default is \code{TRUE}}

\item{constrained}{shall strict monotonicity of constraints be explicitly enforced by linear constraints?}

\item{method}{numerical optimization method}

\item{tol}{tolerance in numerical optimization}

\item{init}{initialization of numerical optimization. Default is neutral}
}
\value{
A list with the following components. 
\describe{
  \item{\code{theahat}}{A vector of estimates for the polychoric correlation coefficient (\code{rho}) as well as thresholds for \code{x} (named \code{a1,a2,...,a_{Kx-1}}) and \code{y} (named \code{b1,b2,...,b_{Ky-1}}).}
  \item{\code{stderr}}{A vector of standard errors for each estimate in \code{theahat}}
  \item{\code{sigma}}{Estimated asymptotic covariance matrix \eqn{\Sigma}, evaluated at the estimates in \code{theahat}.}
  \item{\code{chisq,pval,df}}{Test statistic, p-value, and degrees of freedom of a test for bivariate normality}
  \item{\code{objective}}{Value of minimized loss function}
  \item{\code{optim}}{Object of class \code{optim}}
}
}
\description{
Robust estimation of polychoric correlation coefficient
}
