% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polycor.R
\name{polycor}
\alias{polycor}
\title{Robust estimation of polychoric correlation coefficient}
\usage{
polycor(
  x,
  y,
  c,
  Kx = max(x),
  Ky = max(y),
  variance = TRUE,
  method = "Nelder-Mead",
  tol = 0.001,
  init = c(0, init_thresholds(Kx), init_thresholds(Ky)),
  chisq = FALSE
)
}
\arguments{
\item{x}{vector of integer-valued responses to first item}

\item{y}{vector of integer-valued responses to second item}

\item{c}{tuning constant that governs robustness}

\item{Kx}{number of response options in first item (defaults to \code{max(x)})}

\item{Ky}{number of response options in second item (defaults to \code{max(y)})}

\item{variance}{shall an estimated asymptotic covariance matrix be returned? Default is \code{TRUE}}

\item{method}{numerical optimization method; default is Nelder-Mead}

\item{tol}{tolerance in numerical optimization}

\item{init}{initialization of numerical optimization. Default is neutral}

\item{chisq}{shall a test for bivariate normality of teh latent variables be performed? Default is \code{FALSE}}
}
\description{
Robust estimation of polychoric correlation coefficient
}
