% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polycor.R
\name{polycor_mle}
\alias{polycor_mle}
\title{Maximum likelihood estimation of polychoric correlation coefficient}
\usage{
polycor_mle(
  x,
  y = NULL,
  variance = TRUE,
  constrained = TRUE,
  method = ifelse(constrained, "Nelder-Mead", "L-BFGS-B"),
  tol = 0.001,
  init = initialize_param(x, y)
)
}
\arguments{
\item{x}{vector of integer-valued responses to first item or contingency table (a \code{table} object)}

\item{y}{vector of integer-valued responses to second item; only required if \code{x} is not a contingency table}

\item{variance}{shall an estimated asymptotic covariance matrix be returned? Default is \code{TRUE}}

\item{constrained}{shall strict monotonicity of thresholds be explicitly enforced by linear constraints?}

\item{method}{numerical optimization method; default is Nelder-Mead}

\item{tol}{tolerance in numerical optimization}

\item{init}{initialization of numerical optimization. Default is neutral}
}
\description{
Maximum likelihood estimation of polychoric correlation coefficient
}
