% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.robpolycor}
\alias{print.robpolycor}
\title{Print method for classes \code{"robpolycor"} and \code{"polycor"}.}
\usage{
\method{print}{robpolycor}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{Object of class \code{"robpolycor"} or \code{"polycor"}.}

\item{digits}{Number of digits to be printed.}

\item{...}{Additional parameters to be passed down.}
}
\value{
A print to the console.
}
\description{
Print method for classes \code{"robpolycor"} and \code{"polycor"}.
}
\examples{
set.seed(123)
x <- sample(c(1,2,3), size = 100, replace = TRUE)
y <- sample(c(1,2,3), size = 100, replace = TRUE)
fit <- polycor(x,y) 

print(fit)
fit # equivalent

}
