% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.robpolycor}
\alias{print.robpolycor}
\title{Print method for classes \code{"robpolycor"} and \code{"polycor"}.}
\usage{
\method{print}{robpolycor}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{Object of class \code{"robpolycor"} or \code{"polycor"}.}

\item{digits}{Number of digits to be printed.}

\item{...}{Additional parameters to be passed down.}
}
\description{
Print method for classes \code{"robpolycor"} and \code{"polycor"}.
}
